rename getopt.h to cano_getopt.h
