+++ src/global/getopt.c
@@ -1,4 +1,4 @@
-#include "./getopt.h"
+#include "./cano_getopt.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
